+++ packages/polkadart_scale_codec/analysis_options.yaml
@@ -15,13 +15,12 @@ include: package:lints/recommended.yaml
 
 # Uncomment the following section to specify additional rules.
 
-# linter:
-#   rules:
-#     - camel_case_types
-
+linter:
+  rules:
+    constant_identifier_names: false
+    overridden_fields: false
 # analyzer:
 #   exclude:
-#     - path/to/excluded/files/**
 
 # For more information about the core and recommended set of lints, see
 # https://dart.dev/go/core-lints
