+++ packages/polkadart_scale_codec/lib/polkadart_scale_codec.dart
@@ -1,0 +1,8 @@
+/// Support for doing something awesome.
+///
+/// More dartdocs go here.
+library polkadart_scale_codec;
+
+export 'src/core/polkadart_scale_codec_base.dart';
+
+// Export any libraries intended for clients of this package.
