+++ packages/polkadart_scale_codec/lib/src/core/types.dart
@@ -1,5 +1,3 @@
-
-
 enum Primitive {
   I8,
   U8,
