+++ packages/polkadart_scale_codec/lib/src/util/functions.dart
@@ -1,5 +1,3 @@
-part of scale_codec;
-
 T assertNotNull<T>(T val, String? msg) {
   assert(val != null, msg);
   return val;
