--- packages/scale_codec/CHANGELOG.md
@@ -1,4 +1,0 @@
-## [1.0.0] - 2022-08-24
-
-### First Version
-- First Initialized Version
