--- packages/scale_codec/lib/scale_codec.dart
@@ -1,8 +1,0 @@
-/// Support for doing something awesome.
-///
-/// More dartdocs go here.
-library scale_codec;
-
-export 'src/core/scale_codec_base.dart';
-
-// TODO: Export any libraries intended for clients of this package.
